算法：点到任意维度的直线、线段的距离 算法：二维平面中点集合中到直线的最大距离的点

+++ src/gww/geeks/geometry/Point.java
@@ -6,8 +6,8 @@ import java.util.Comparator;
  * Created by 高文文 on 2017/5/10.
  */
 public class Point implements Comparator<Point>{
-    float x, y;
-    public Point(float x, float y) {
+    double x, y;
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
